@@ -4,13 +4,13 @@ description: A fluid bottom navigation bar example
 version: 1.1.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
   fluid_bottom_nav_bar:
     path: ../
