@@ -4,13 +4,12 @@ version: 1.2.0
 homepage: https://github.com/eric-taix/fluid-nav-bar
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=1.20.1 <2.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^0.18.0
 
 dev_dependencies:
   flutter_test:
