@@ -7,8 +7,7 @@ import './fluid_nav_bar_item.dart';
 
 typedef void FluidNavBarChangeCallback(int selectedIndex);
 
-typedef Widget FluidNavBarItemBuilder(
-    FluidNavBarIcon icon, FluidNavBarItem item);
+typedef Widget FluidNavBarItemBuilder(FluidNavBarIcon icon, FluidNavBarItem item);
 
 /// A widget to display a fluid navigation bar with icon buttons.
 ///
@@ -34,10 +33,10 @@ class FluidNavBar extends StatefulWidget {
   final List<FluidNavBarIcon> icons;
 
   /// A callback called when an icon has been tapped with its index
-  final FluidNavBarChangeCallback onChange;
+  final FluidNavBarChangeCallback? onChange;
 
   /// The style to use to paint the fluid navigation bar and its icons
-  final FluidNavBarStyle style;
+  final FluidNavBarStyle? style;
 
   /// Delay to adjust the overall delay of the animations
   ///   * < 1 is faster
@@ -47,7 +46,7 @@ class FluidNavBar extends StatefulWidget {
 
   /// The scale factor used when an icon is tapped
   /// 1.0 means that the icon is not scaled and 1.5 means the icons is scaled to +50%
-  /// An optional builder to change or wrap the builded item
+  /// An optional builder to change or wrap the built item
   ///
   /// This is where you can wrap the item with semantic or
   /// other widget
@@ -60,31 +59,29 @@ class FluidNavBar extends StatefulWidget {
   final FluidNavBarItemBuilder itemBuilder;
 
   FluidNavBar(
-      {Key key,
-      @required this.icons,
+      {Key? key,
+      required this.icons,
       this.onChange,
       this.style,
       this.animationFactor = 1.0,
       this.scaleFactor = 1.2,
       this.defaultIndex = 0,
-      FluidNavBarItemBuilder itemBuilder})
+      FluidNavBarItemBuilder? itemBuilder})
       : this.itemBuilder = itemBuilder ?? _identityBuilder,
-        assert(icons != null && icons.length > 1),
+        assert(icons.length > 1),
         super(key: key);
 
   @override
   State createState() => _FluidNavBarState();
 
-  static Widget _identityBuilder(FluidNavBarIcon icon, FluidNavBarItem item) =>
-      item;
+  static Widget _identityBuilder(FluidNavBarIcon icon, FluidNavBarItem item) => item;
 }
 
-class _FluidNavBarState extends State<FluidNavBar>
-    with TickerProviderStateMixin {
+class _FluidNavBarState extends State<FluidNavBar> with TickerProviderStateMixin {
   int _currentIndex = 0;
 
-  AnimationController _xController;
-  AnimationController _yController;
+  AnimationController? _xController;
+  AnimationController? _yController;
 
   @override
   void initState() {
@@ -92,10 +89,8 @@ class _FluidNavBarState extends State<FluidNavBar>
 
     _currentIndex = widget.defaultIndex;
 
-    _xController = AnimationController(
-        vsync: this, animationBehavior: AnimationBehavior.preserve);
-    _yController = AnimationController(
-        vsync: this, animationBehavior: AnimationBehavior.preserve);
+    _xController = AnimationController(vsync: this, animationBehavior: AnimationBehavior.preserve);
+    _yController = AnimationController(vsync: this, animationBehavior: AnimationBehavior.preserve);
 
     Listenable.merge([_xController, _yController]).addListener(() {
       setState(() {});
@@ -104,17 +99,16 @@ class _FluidNavBarState extends State<FluidNavBar>
 
   @override
   void didChangeDependencies() {
-    _xController.value =
-        _indexToPosition(_currentIndex) / MediaQuery.of(context).size.width;
-    _yController.value = 1.0;
+    _xController!.value = _indexToPosition(_currentIndex) / MediaQuery.of(context).size.width;
+    _yController!.value = 1.0;
 
     super.didChangeDependencies();
   }
 
   @override
   void dispose() {
-    _xController.dispose();
-    _yController.dispose();
+    _xController!.dispose();
+    _yController!.dispose();
     super.dispose();
   }
 
@@ -140,9 +134,7 @@ class _FluidNavBarState extends State<FluidNavBar>
             top: 0,
             width: _getButtonContainerWidth(),
             height: height,
-            child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: _buildButtons()),
+            child: Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: _buildButtons()),
           ),
         ],
       ),
@@ -152,12 +144,12 @@ class _FluidNavBarState extends State<FluidNavBar>
   Widget _buildBackground() {
     return CustomPaint(
       painter: _BackgroundCurvePainter(
-        _xController.value * MediaQuery.of(context).size.width,
+        _xController!.value * MediaQuery.of(context).size.width,
         Tween<double>(
-          begin: Curves.easeInExpo.transform(_yController.value),
-          end: ElasticOutCurve(0.38).transform(_yController.value),
-        ).transform(_yController.velocity.sign * 0.5 + 0.5),
-        widget?.style?.barBackgroundColor ?? Colors.white,
+          begin: Curves.easeInExpo.transform(_yController!.value),
+          end: ElasticOutCurve(0.38).transform(_yController!.value),
+        ).transform(_yController!.velocity.sign * 0.5 + 0.5),
+        widget.style?.barBackgroundColor ?? Colors.white,
       ),
     );
   }
@@ -170,19 +162,14 @@ class _FluidNavBarState extends State<FluidNavBar>
           (entry) => widget.itemBuilder(
             entry.value,
             FluidNavBarItem(
-              entry.value.svgPath ?? entry.value.iconPath,
               entry.value.icon,
               _currentIndex == entry.key,
               () => _handleTap(entry.key),
-              entry.value.selectedForegroundColor ??
-                  widget?.style?.iconSelectedForegroundColor ??
-                  Colors.black,
-              entry.value.unselectedForegroundColor ??
-                  widget?.style?.iconUnselectedForegroundColor ??
-                  Colors.grey,
+              entry.value.selectedForegroundColor ?? widget.style?.iconSelectedForegroundColor ?? Colors.black,
+              entry.value.unselectedForegroundColor ?? widget.style?.iconUnselectedForegroundColor ?? Colors.grey,
               entry.value.backgroundColor ??
-                  widget?.style?.iconBackgroundColor ??
-                  widget?.style?.barBackgroundColor ??
+                  widget.style?.iconBackgroundColor ??
+                  widget.style?.barBackgroundColor ??
                   Colors.white,
               widget.scaleFactor,
               widget.animationFactor,
@@ -207,34 +194,29 @@ class _FluidNavBarState extends State<FluidNavBar>
     final appWidth = MediaQuery.of(context).size.width;
     final buttonsWidth = _getButtonContainerWidth();
     final startX = (appWidth - buttonsWidth) / 2;
-    return startX +
-        index.toDouble() * buttonsWidth / buttonCount +
-        buttonsWidth / (buttonCount * 2.0);
+    return startX + index.toDouble() * buttonsWidth / buttonCount + buttonsWidth / (buttonCount * 2.0);
   }
 
   void _handleTap(int index) {
-    if (_currentIndex == index || _xController.isAnimating) return;
+    if (_currentIndex == index || _xController!.isAnimating) return;
 
     setState(() {
       _currentIndex = index;
     });
 
-    _yController.value = 1.0;
-    _xController.animateTo(
-        _indexToPosition(index) / MediaQuery.of(context).size.width,
+    _yController!.value = 1.0;
+    _xController!.animateTo(_indexToPosition(index) / MediaQuery.of(context).size.width,
         duration: Duration(milliseconds: 620) * widget.animationFactor);
     Future.delayed(
       Duration(milliseconds: 500) * widget.animationFactor,
       () {
-        _yController.animateTo(1.0,
-            duration: Duration(milliseconds: 1200) * widget.animationFactor);
+        _yController!.animateTo(1.0, duration: Duration(milliseconds: 1200) * widget.animationFactor);
       },
     );
-    _yController.animateTo(0.0,
-        duration: Duration(milliseconds: 300) * widget.animationFactor);
+    _yController!.animateTo(0.0, duration: Duration(milliseconds: 300) * widget.animationFactor);
 
     if (widget.onChange != null) {
-      widget.onChange(index);
+      widget.onChange!(index);
     }
   }
 }
@@ -268,32 +250,26 @@ class _BackgroundCurvePainter extends CustomPainter {
     // Paint two cubic bezier curves using various linear interpolations based off of the `_normalizedY` value
     final norm = LinearPointCurve(0.5, 2.0).transform(_normalizedY) / 2;
 
-    final radius =
-        Tween<double>(begin: _radiusTop, end: _radiusBottom).transform(norm);
+    final radius = Tween<double>(begin: _radiusTop, end: _radiusBottom).transform(norm);
     // Point colinear to the top edge of the background pane
-    final anchorControlOffset = Tween<double>(
-            begin: radius * _horizontalControlTop,
-            end: radius * _horizontalControlBottom)
-        .transform(LinearPointCurve(0.5, 0.75).transform(norm));
+    final anchorControlOffset =
+        Tween<double>(begin: radius * _horizontalControlTop, end: radius * _horizontalControlBottom)
+            .transform(LinearPointCurve(0.5, 0.75).transform(norm));
     // Point that slides up and down depending on distance for the target x position
-    final dipControlOffset = Tween<double>(
-            begin: radius * _pointControlTop, end: radius * _pointControlBottom)
+    final dipControlOffset = Tween<double>(begin: radius * _pointControlTop, end: radius * _pointControlBottom)
         .transform(LinearPointCurve(0.5, 0.8).transform(norm));
-    final y = Tween<double>(begin: _topY, end: _bottomY)
-        .transform(LinearPointCurve(0.2, 0.7).transform(norm));
-    final dist = Tween<double>(begin: _topDistance, end: _bottomDistance)
-        .transform(LinearPointCurve(0.5, 0.0).transform(norm));
+    final y = Tween<double>(begin: _topY, end: _bottomY).transform(LinearPointCurve(0.2, 0.7).transform(norm));
+    final dist =
+        Tween<double>(begin: _topDistance, end: _bottomDistance).transform(LinearPointCurve(0.5, 0.0).transform(norm));
     final x0 = _x - dist / 2;
     final x1 = _x + dist / 2;
 
     final path = Path()
       ..moveTo(0, 0)
       ..lineTo(x0 - radius, 0)
-      ..cubicTo(
-          x0 - radius + anchorControlOffset, 0, x0 - dipControlOffset, y, x0, y)
+      ..cubicTo(x0 - radius + anchorControlOffset, 0, x0 - dipControlOffset, y, x0, y)
       ..lineTo(x1, y)
-      ..cubicTo(x1 + dipControlOffset, y, x1 + radius - anchorControlOffset, 0,
-          x1 + radius, 0)
+      ..cubicTo(x1 + dipControlOffset, y, x1 + radius - anchorControlOffset, 0, x1 + radius, 0)
       ..lineTo(size.width, 0)
       ..lineTo(size.width, size.height)
       ..lineTo(0, size.height);
@@ -305,8 +281,6 @@ class _BackgroundCurvePainter extends CustomPainter {
 
   @override
   bool shouldRepaint(_BackgroundCurvePainter oldPainter) {
-    return _x != oldPainter._x ||
-        _normalizedY != oldPainter._normalizedY ||
-        _color != oldPainter._color;
+    return _x != oldPainter._x || _normalizedY != oldPainter._normalizedY || _color != oldPainter._color;
   }
 }
