@@ -2,7 +2,6 @@ import 'package:fluid_bottom_nav_bar/fluid_bottom_nav_bar.dart';
 import 'package:fluid_bottom_nav_bar/src/fluid_nav_bar_icon.dart';
 import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
 
 import 'curves.dart';
 
@@ -21,11 +20,8 @@ typedef void FluidNavBarButtonTappedCallback();
 class FluidNavBarItem extends StatefulWidget {
   static const nominalExtent = const Size(64, 64);
 
-  /// The path of the SVG asset
-  final String svgPath;
-
   // The icon data
-  final IconData icon;
+  final IconData? icon;
 
   /// Flag to know if this item is active or not
   final bool selected;
@@ -49,7 +45,6 @@ class FluidNavBarItem extends StatefulWidget {
   final double animationFactor;
 
   FluidNavBarItem(
-    this.svgPath,
     this.icon,
     this.selected,
     this.onTap,
@@ -58,11 +53,7 @@ class FluidNavBarItem extends StatefulWidget {
     this.backgroundColor,
     this.scaleFactor,
     this.animationFactor,
-  )   : assert(scaleFactor >= 1.0),
-        assert(svgPath == null || icon == null,
-            'Cannot provide both an iconPath and an icon.'),
-        assert(!(svgPath == null && icon == null),
-            'An iconPath or an icon must be provided.');
+  ) : assert(scaleFactor >= 1.0);
 
   @override
   State createState() {
@@ -70,19 +61,18 @@ class FluidNavBarItem extends StatefulWidget {
   }
 }
 
-class _FluidNavBarItemState extends State<FluidNavBarItem>
-    with SingleTickerProviderStateMixin {
+class _FluidNavBarItemState extends State<FluidNavBarItem> with SingleTickerProviderStateMixin {
   static const double _activeOffset = 16;
   static const double _defaultOffset = 0;
   static const double _iconSize = 25;
 
   bool _selected;
 
-  AnimationController _animationController;
-  Animation<double> _activeColorClipAnimation;
-  Animation<double> _yOffsetAnimation;
-  Animation<double> _activatingAnimation;
-  Animation<double> _inactivatingAnimation;
+  late AnimationController _animationController;
+  late Animation<double> _activeColorClipAnimation;
+  late Animation<double> _yOffsetAnimation;
+  late Animation<double> _activatingAnimation;
+  late Animation<double> _inactivatingAnimation;
 
   _FluidNavBarItemState(this._selected);
 
@@ -93,23 +83,19 @@ class _FluidNavBarItemState extends State<FluidNavBarItem>
     double waveRatio = 0.28;
     _animationController = AnimationController(
       duration: Duration(milliseconds: (1600 * widget.animationFactor).toInt()),
-      reverseDuration:
-          Duration(milliseconds: (1000 * widget.animationFactor).toInt()),
+      reverseDuration: Duration(milliseconds: (1000 * widget.animationFactor).toInt()),
       vsync: this,
     )..addListener(() => setState(() {}));
 
-    _activeColorClipAnimation =
-        Tween<double>(begin: 0.0, end: _iconSize).animate(CurvedAnimation(
+    _activeColorClipAnimation = Tween<double>(begin: 0.0, end: _iconSize).animate(CurvedAnimation(
       parent: _animationController,
       curve: Interval(0.25, 0.38, curve: Curves.easeOut),
       reverseCurve: Interval(0.7, 1.0, curve: Curves.easeInCirc),
     ));
 
-    var _animation = CurvedAnimation(
-        parent: _animationController, curve: LinearPointCurve(waveRatio, 0.0));
+    var _animation = CurvedAnimation(parent: _animationController, curve: LinearPointCurve(waveRatio, 0.0));
 
-    _yOffsetAnimation = Tween<double>(begin: _defaultOffset, end: _activeOffset)
-        .animate(CurvedAnimation(
+    _yOffsetAnimation = Tween<double>(begin: _defaultOffset, end: _activeOffset).animate(CurvedAnimation(
       parent: _animation,
       curve: ElasticOutCurve(0.38),
       reverseCurve: Curves.easeInCirc,
@@ -118,8 +104,7 @@ class _FluidNavBarItemState extends State<FluidNavBarItem>
     var activatingHalfTween = Tween<double>(begin: 1, end: widget.scaleFactor);
     _activatingAnimation = TweenSequence([
       TweenSequenceItem(tween: activatingHalfTween, weight: 50.0),
-      TweenSequenceItem(
-          tween: ReverseTween<double>(activatingHalfTween), weight: 50.0),
+      TweenSequenceItem(tween: ReverseTween<double>(activatingHalfTween), weight: 50.0),
     ]).animate(CurvedAnimation(
       parent: _animation,
       curve: Interval(0.0, 0.3),
@@ -153,8 +138,7 @@ class _FluidNavBarItemState extends State<FluidNavBarItem>
   Widget build(context) {
     const ne = FluidNavBarItem.nominalExtent;
 
-    final scaleAnimation =
-        _selected ? _activatingAnimation : _inactivatingAnimation;
+    final scaleAnimation = _selected ? _activatingAnimation : _inactivatingAnimation;
 
     return GestureDetector(
       onTap: widget.onTap,
@@ -173,38 +157,21 @@ class _FluidNavBarItemState extends State<FluidNavBarItem>
           child: Stack(children: <Widget>[
             Container(
               alignment: Alignment.center,
-              child: widget.icon == null
-                  ? SvgPicture.asset(
-                      widget.svgPath,
-                      color: widget.unselectedForegroundColor,
-                      width: _iconSize,
-                      height: _iconSize * scaleAnimation.value,
-                      colorBlendMode: BlendMode.srcIn,
-                    )
-                  : Icon(
-                      widget.icon,
-                      color: widget.unselectedForegroundColor,
-                      size: _iconSize * scaleAnimation.value,
-                    ),
+              child: Icon(
+                widget.icon,
+                color: widget.unselectedForegroundColor,
+                size: _iconSize * scaleAnimation.value,
+              ),
             ),
             Container(
               alignment: Alignment.center,
               child: ClipRect(
-                clipper: _SvgPictureClipper(
-                    _activeColorClipAnimation.value * scaleAnimation.value),
-                child: widget.icon == null
-                    ? SvgPicture.asset(
-                        widget.svgPath,
-                        color: widget.selectedForegroundColor,
-                        width: _iconSize,
-                        height: _iconSize * scaleAnimation.value,
-                        colorBlendMode: BlendMode.srcIn,
-                      )
-                    : Icon(
-                        widget.icon,
-                        color: widget.selectedForegroundColor,
-                        size: _iconSize * scaleAnimation.value,
-                      ),
+                clipper: _SvgPictureClipper(_activeColorClipAnimation.value * scaleAnimation.value),
+                child: Icon(
+                  widget.icon,
+                  color: widget.selectedForegroundColor,
+                  size: _iconSize * scaleAnimation.value,
+                ),
               ),
             ),
           ]),
@@ -232,8 +199,7 @@ class _SvgPictureClipper extends CustomClipper<Rect> {
 
   @override
   Rect getClip(Size size) {
-    return Rect.fromPoints(size.topLeft(Offset.zero),
-        size.topRight(Offset.zero) + Offset(0, height));
+    return Rect.fromPoints(size.topLeft(Offset.zero), size.topRight(Offset.zero) + Offset(0, height));
   }
 
   @override
